Game gear refactors
